Preload parameter and subtitle supports.
The first change proposal is to switch the default from `preload="auto"` to `preload="metadata"` and to let the parameter change directly from the shortcode call. Depending on the type of hosting a video can cause huge data bills (server side and user side) when it preloads almost [50s of video ](https://www.stevesouders.com/blog/2013/04/12/html5-video-preload/) with the `auto` parameter.

Second proposal was to add support for subtitles in VTT format.  The shortcode will load the subtitles automatically from the page bundle as long as you name them as `[video-filename].[language_code].vtt`. The subtitle should follow the [VTT format](https://developer.mozilla.org/en-US/docs/Web/API/WebVTT_API) and the language code must be 2 letters. 
It follows the same logic as the poster file so it shouldn't be too confusing.